mm: thp: cleanup and optimize compound_nr
Minor cleanup and minor optimizations to compound_nr and its callers.

There's no reason to return "unsigned long" in compound_nr, when it's
defined as "unsigned int" in the page structure and it's occasionally
truncated back to a 32bit value in its callers.

In addition its retval should be allowed to be cached in a caller
saved register and not be forced to be re-read from memory at every
iteration if any function called in the loop isn't in scope or inline.

+++ include/linux/mm.h
@@ -949,7 +949,7 @@ static inline void set_compound_order(struct page *page, unsigned int order)
 }
 
 /* Returns the number of pages in this potentially compound page. */
-static inline unsigned long compound_nr(struct page *page)
+static inline unsigned int compound_nr(struct page *page)
 {
 	if (!PageHead(page))
 		return 1;
